+++ demo/src/main/java/com/orange/ods/demo/ui/components/buttons/ButtonsContained.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import com.orange.ods.compose.component.OdsComponent
 import com.orange.ods.compose.component.button.OdsButton
 import com.orange.ods.compose.component.button.OdsButtonStyle
 import com.orange.ods.compose.component.chip.OdsChoiceChip
@@ -33,8 +34,11 @@ import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.theme.OdsDisplaySurface
 import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
+import com.orange.ods.demo.ui.utilities.composable.ButtonTechnicalText
+import com.orange.ods.demo.ui.utilities.composable.CodeImplementationColumn
 import com.orange.ods.demo.ui.utilities.composable.Subtitle
 import com.orange.ods.demo.ui.utilities.composable.SwitchListItem
+import com.orange.ods.demo.ui.utilities.composable.Title
 import com.orange.ods.utilities.extension.fullName
 
 @OptIn(ExperimentalMaterialApi::class)
@@ -46,16 +50,6 @@ fun ButtonsContained(style: OdsButtonStyle) {
         ComponentCustomizationBottomSheetScaffold(
             bottomSheetScaffoldState = rememberBottomSheetScaffoldState(),
             bottomSheetContent = {
-                Subtitle(textRes = R.string.component_style, horizontalPadding = true)
-                OdsChoiceChipsFlowRow(
-                    selectedChip = containedButtonStyle,
-                    modifier = Modifier.padding(horizontal = dimensionResource(id = R.dimen.screen_horizontal_margin)),
-                    outlinedChips = true
-                ) {
-                    OdsChoiceChip(textRes = R.string.component_button_style_primary, value = OdsButtonStyle.Primary)
-                    OdsChoiceChip(textRes = R.string.component_button_style_default, value = OdsButtonStyle.Default)
-                    OdsChoiceChip(textRes = R.string.component_button_style_functional, value = OdsButtonStyle.FunctionalPositive)
-                }
                 if (containedButtonStyle.value in listOf(OdsButtonStyle.FunctionalPositive, OdsButtonStyle.FunctionalNegative)) {
                     Subtitle(textRes = R.string.component_button_style_functional, horizontalPadding = true)
                     OdsChoiceChipsFlowRow(
@@ -77,7 +71,14 @@ fun ButtonsContained(style: OdsButtonStyle) {
                     .verticalScroll(rememberScrollState())
                     .padding(vertical = dimensionResource(id = R.dimen.screen_vertical_margin))
             ) {
-                ButtonStyleTitle(style = containedButtonStyle.value)
+                with(containedButtonStyle.value) {
+                    if (this in listOf(OdsButtonStyle.FunctionalNegative, OdsButtonStyle.FunctionalPositive)) {
+                        Title(
+                            textRes = if (this == OdsButtonStyle.FunctionalNegative) R.string.component_button_style_functional_negative else R.string.component_button_style_functional_positive,
+                            horizontalPadding = true
+                        )
+                    }
+                }
 
                 ContainedButton(style = containedButtonStyle.value, leadingIcon = hasLeadingIcon, enabled = isEnabled, variableWidth = hasVariableWidth)
 
@@ -104,35 +105,19 @@ fun ButtonsContained(style: OdsButtonStyle) {
                         )
                     }
                 }
+
+                CodeImplementationColumn {
+                    ButtonTechnicalText(
+                        componentName = OdsComponent.OdsButton.name,
+                        style = containedButtonStyle.value.fullName,
+                        enabled = isEnabled,
+                        variableWidth = hasVariableWidth,
+                        icon = hasLeadingIcon
+                    )
+                }
             }
         }
     }
-}
-
-@Composable
-private fun ButtonStyleTitle(style: OdsButtonStyle) {
-    val titleRes: Int
-    val technicalText: String
-    when (style) {
-        OdsButtonStyle.Default -> {
-            titleRes = R.string.component_button_style_default
-            technicalText = OdsButtonStyle.Default.fullName
-        }
-        OdsButtonStyle.Primary -> {
-            titleRes = R.string.component_button_style_primary
-            technicalText = OdsButtonStyle.Primary.fullName
-        }
-        OdsButtonStyle.FunctionalNegative -> {
-            titleRes = R.string.component_button_style_functional_negative
-            technicalText = OdsButtonStyle.FunctionalNegative.fullName
-        }
-        OdsButtonStyle.FunctionalPositive -> {
-            titleRes = R.string.component_button_style_functional_positive
-            technicalText = OdsButtonStyle.FunctionalPositive.fullName
-        }
-    }
-
-    StyleTitle(titleRes = titleRes, technicalText = technicalText)
 }
 
 @Composable
@@ -143,9 +128,7 @@ private fun ContainedButton(
     variableWidth: Boolean,
     displaySurface: OdsDisplaySurface = OdsDisplaySurface.Default
 ) {
-    val modifier = Modifier
-        .padding(horizontal = dimensionResource(R.dimen.screen_horizontal_margin))
-        .padding(top = dimensionResource(R.dimen.spacing_m))
+    val modifier = Modifier.padding(horizontal = dimensionResource(R.dimen.screen_horizontal_margin), vertical = dimensionResource(R.dimen.spacing_m))
 
     OdsButton(
         modifier = if (variableWidth) modifier else modifier.fillMaxWidth(),